add delete route, not work, CSRF token not passed properly

+++ app/api/delete-file/route.ts
@@ -1,15 +1,21 @@
 // app/api/delete-file/route.ts
 
+import { auth } from '@/auth';
 import { NextRequest, NextResponse } from 'next/server';
 
 export async function DELETE(req: NextRequest) {
+  const session = await auth();
+if (!session) {
+  console.error("Unauthorized request to delete file.");
+  return false;
+}
   try {
-    const { route } = await req.json();
+    const { fileRoute } = await req.json();
 
-    const url = `https://api.imagekit.io/v1/${route}`;
+    const url = `https://api.imagekit.io/v1/${fileRoute}`;
     const privateKey = process.env.IMAGEKIT_PRIVATE_KEY!;
     const base64Auth = Buffer.from(privateKey + ":").toString("base64");
-
+console.log("Deleting file with route:", base64Auth, url);
     const options = {
       method: 'DELETE',
       headers: {
@@ -21,6 +27,7 @@ export async function DELETE(req: NextRequest) {
     const response = await fetch(url, options);
     const data = await response.json();
        if (!response.ok) {
+        console.error("Failed to delete file:", data);
       return NextResponse.json({ success: false, error: data.error }, { status: response.status });
     }
